# Conflicts:
#	t1g-distro/general/ddls/local_api_engine_postgres_inserts.dml
#	t1g-distro/general/ddls/t1t_acc_api_engine_postgres_inserts.dml
#	t1g-distro/general/ddls/t1t_dev_api_engine_postgres_inserts.dml
#	t1g-distro/general/ddls/t1t_prod_api_engine_postgres_inserts.dml

Updated font-awesome icon for HAL & DataDog

+++ t1g-ejb/src/main/java/com/t1t/apim/gateway/rest/RestGatewayLink.java
@@ -230,7 +230,7 @@ public class RestGatewayLink implements IGatewayLink {
     @Override
     public Integer getCentralOAuthTokenExpirationTime() throws GatewayAuthenticationException {
         final KongPluginOAuthEnhanced oAuth2ExpirationForCentralOAuth = getClient().getOAuth2ExpirationForCentralOAuth(gateway);
-        return oAuth2ExpirationForCentralOAuth.getTokenExpiration();
+        return oAuth2ExpirationForCentralOAuth.getTokenExpiration().intValue();
     }
 
     /**
